+++ server_run.py
@@ -1,0 +1,6 @@
+import server
+
+new_server = server.Server()
+new_server.connection()
+new_server.game()
+new_server.close_connection()
